Module odilia::cache

Structs

AccessiblePrimitive

A struct which represents the bare minimum of an accessible for purposes of caching. This makes some possibly eronious assumptions about what the sender is. TODO: find a better way to store this info where it still imeplements Copy.

Cache

The root of the accessible cache.

CacheItem

A struct representing an accessible. To get any information from the cache other than the stored information like role, interfaces, and states, you will need to instantiate an atspi::accessible::AccessibleProxy or other *Proxy type from atspi to query further info.

Functions

copy_into_cache_item 🔒

Copy all info into a plain CacheItem struct. This is very cheap, and the locking overhead will vastly outstrip making this a non-copy struct.

Type Definitions

FxBuildHasher 🔒

FxReadGuard

FxReadHandleFactory

FxWriteHandle