Function odilia::events::object::text_changed::live_to_priority

pub fn live_to_priority(live_str: &AriaLive) -> Priority

if the aria-live attribute is set to “polite”, then set the priority of the message to speak once all other messages are done if the aria-live attribute is set to “assertive”, then set the priority of the message to speak immediately, stop all other messages, and do not interrupt that piece of speech otherwise, do not continue