odilia

Module tower

Re-exports

pub use handler::Handler;

pub use service_ext::ServiceExt;

pub use extractors::*;

pub use handlers::*;

Modules

async_try

A trait AsyncTryFrom, its associated layer AsyncTryIntoLayer, and a blanket implementation of odilia_common::from_state::TryFromState.

choice

extractors

from_state

handler

handlers

iter_svc

predicate 🔒

service_ext

service_set

state_changed

state_svc

A generic state provider service. This clones the state upon every invocation of Service::call, so make sure it’s relatively cheap to do so.

sync_try

Synchronous version of crate::async_try. Create tower::Services out of the ability to convert values via TryFrom and TryInto].

unwrap_svc

A service built specifically to run .unwrap() on computed futures.