odilia::tower

Module async_try

A trait AsyncTryFrom, its associated layer AsyncTryIntoLayer, and a blanket implementation of odilia_common::from_state::TryFromState.

Due to the blanket implementation, this is required to be a trait defined by us. This means that even in the future, if a crate were to become available that offered similar functionality, we could still not remove this.

Structs

AndThenCall

A version of tower::util::future::AndThenFuture that is not generic over an un-namable future type.

AsyncTryIntoLayer

A ZST representing a AsyncTryInto as a tower::Layer.

AsyncTryIntoService

A service which applies an AsyncTryInto transformation to a service’s input.

Traits

AsyncTryFrom

An async version of TryFrom with an associated future.

AsyncTryInto

An async version of TryInto with an associated future.