odilia::tower

Module state_svc

A generic state provider service. This clones the state upon every invocation of Service::call, so make sure it’s relatively cheap to do so.

Structs

StateLayer

A tower::Layer which stores state S.

StateService

A service which clones state Sta into the Service::call method of the given service Srv.