Module odilia::events::object::text_changed

Functions

append_to_object

dispatch

get_atomic_state

get_live_state

Get the live state of a set of attributes. Although the function only currently tests one attributes, in the future it may be imporant to inspect many attributes, compare them, or do additional logic.

get_string_within_bounds

get_string_without_bounds

insert_at_index

insert_at_range

insert_or_delete

The insert boolean, if set to true, will update the text in the cache. If it is set to false, the selection will be removed. The TextChangedEvent::kind value will NOT be checked by this function.

live_to_priority

if the aria-live attribute is set to “polite”, then set the prioirty of the message to speak once all other messages are done if the aria-live attribute is set to “assertive”, thenset the priority of the message to speak immediately, stop all other messages, and do not interrupt that piece of speech otherwise, do not continue

speak_insertion

update_string_insert